app_dialtone: add wait for digit functionality
Behavior is changed so that if a negative number appears in the digit map,
it will now wait the absolute value in seconds and return if nothing has
been entered. This is useful for options like reading a 0 and allowing 0+
dialing, but completing quickly on 0- calls with a small timeout.