func_query: Add TEXT_QUERY function
Adds the TEXT_QUERY function, which makes it easy for
users to be able to remotely query data available on
remote systems.

For example, the result of functions like DEVICE_STATE
can be made easily available to other systems through
this function, which presents such data as easily as if
it were available locally. As it is tech-agnostic, behind
the scenes, a call is initiated to the specified endpoint,
and then the buffer is filled with the results of the
data transfer.

Data transfer remains done using SendText; this function
merely makes it seamless to use such data from other systems.