res_irc: fix long messages causing client disconnect
Because IRC_BUFFER_SIZE was 256, if a message longer than
256 characters was received, this would fill up the entire
available buffer. We would realize we hadn't received the full
message yet, due to the absence of CR, but calling read
again would return 0 because the actual buffer size available
was 0.

This is immediately fixed by flushing the contents of the buffer
if it actually fills all the way up so that input can be read back
into an empty buffer.

Additionally, IRC messages have a maximum length of 512, not
256, so the buffer size itself is increased to 512.

This will prevent this scenario from happening at all on well-formatted
messages, but even if we fill up the entire buffer for some reason, this
will no longer cause read to return 0.