res_coindetect: add COIN_EIS function for Expanding In-Band signaling
Adds a COIN_EIS function that can be used for Expanded In-Band signaling
support on the slave end of a Feature Group C EIS trunk. This greatly simplifies
EIS integration with the dialplan as the wink and MF digit can be detected without
needing to fork Dial for bridging, which means that correct answer supervision
can pass thru to the Class 5 caller.