0.1.94: add keyperms command
Add command to ensure that TLS keys are sufficiently readable.
If TLS keys are used by certain Asterisk modules or components and
they are not properly readable (not just the keys but all their parent directories),
this can cause serious issues. Therefore, we now have logic to be able to check
keys that are used and make sure that they are sufficiently readable.