patches: Don't change EXTEN for prefix includes.
If we change the EXTEN, that will cause dialplan execution to fail if there
is more than 1 priority, so we no longer do this.
Changing the context as well is another hack that theoretically works, since
at that point prefix includes are out of the picture, but this really breaks
the semantics of how includes in the dialplan are supposed to work, so I
opted not to do this.
This does mean that in contexts that could be the target of a prefixed include,
the ${EXTEN} special variable should not be used literally, unless, of course,
it is only included as a prefix include and never as a regular include. This is
because ${EXTEN} will be the original extension, including the original prefix,
so usage of ${EXTEN} will likely result in incorrect behavior. Instead, a substring
of the last N characters should be taken, e.g. ${EXTEN:-10}. ${EXTEN:1} would not
be acceptable, since this assumes that we used a prefixed include, but if we didn't,
that would strip a digit it shouldn't have.
Furthermore, for this reason patterns ending in . should not be used, since the
length of the pattern essentially needs to be hardcoded into any uses of ${EXTEN}
in any context that could be called from a prefix include.
Admittedly, none of the approaches discussed here are 100% elegant, but this is
likely a necessary workaround due to the semihack that prefixed includes themselves are.