app_callerid: Add off-hook Caller ID application.
Adds the (partially working) SendCWCID application.
This application was written earlier in 2022 and I have
not been able to return to it since non-DAHDI handling
still needs to be fixed, and appears to suffer from the
same issue that caused ADSI to break starting in v13.

The application currently works properly on DAHDI channels.
It does not currently work on non-DAHDI channels.

That will need to be fixed down the line, but I'm committing
the module in its current form now since what works is
stable and not likely to change for a while.