app_loopdisconnect: Add LoopDisconnect application.
This application was supposed to be added this spring,
but I never did that and then I lost the code for it, so
I rewrote the patches and application for it.

Because this requires DAHDI changes, it can't be
merged into Asterisk anytime soon, since the DAHDI
team is (still) AWOL, so the LoopDisconnect application
is in its own module, just like app_flash, to avoid
adding to the already large chan_dahdi module and to
minimize the number of conflicts in chan_dahdi.c.