app_george: Add George, the Interactive Answering Machine.
Adds an application that implements George, the Interactive
Answering Machine, a cool interactive telephone answering
device that Evan Doorbell invented in the early 1980s.

This application has been done for a while, I just cleaned
it up a little today and added a config file so prompt files
aren't hardcoded into the module.