blueboxing.diff: Fix SF receiver regression
Unfortunately, incrementing hit_count whenever there is any hit at all, more specifically on the first one, led to a regression with the SF receiver not working properly. For now, this is reverted and in the future this may be revisited if we can find a non-breaking way to make this change.