install: Trust invalid SSL cert for MP3 source.
The MP3 source is in SVN, and it's now happened
several times that the certificate is not trusted
when attempting to run an SVN command against
the Digium SVN server. This is documented at
https://issues.asterisk.org/jira/browse/ASTERISK-30170

This will cause installs to fail if something
requiring the MP3 source is selected for installation.

As a temporary workaround, always trust the certificate
and download manually before running the MP3 script,
so that we can properly download and patch the MP3 files.