app_softmodem: Properly write entire buffer to TDD modem.
v18_put returns the number of bytes actually written to
the modem successfully, and if this number is less than
the number of bytes we attempted to write, then not all
the data will actually get sent by the V.18 modem, so
we need to keep retrying to send the remainder of the contents
of the buffer. This is now done, and truncation of long
messages sent to a TDD no longer occurs.