coredumps: Fix GDB installation detection.
Some systems seem to have a "stub" for GDB
which will result in the following error
if trying to use it when not really installed:
/usr/bin/gdb does not support python

The solution is to simply go ahead and install
GDB. However, we're not catching this because
if we run "which gdb", it seems to exist, from
that perspective. Therefore, we're missing an
opportunity to automatically install it.

The GDB installation check is now enhanced a
little bit to hopefully avoid being fooled by
this in the future. We don't want to blindly
install it every time since if it's already
installed, it will still waste time running
apt commands.