patches: Remove gerrit commands and references.
Gerrit has not been used for the Asterisk project since
the end of April and, as of this month, is no longer available,
so any remaining Gerrit references are effectively no longer
useful. This (now useless) code is removed. Additionally,
a command is added for the new post-migration documentation
generator.