res_phreaknet: Add warning for IAX2 dialing bugs.
It seems that on systems where calls to a hostname which
resolves locally to the loopback address will fail
with IAX2. This is outlined here: https://github.com/asterisk/asterisk/issues/413

This might happen on nodes where a (public) hostname
resolves internally to a loopback IP address, but chan_iax2
is not listening on that specific address (e.g. 127.0.1.1
instead of 127.0.0.1). This can likely be resolved by adjusting
/etc/hosts.