app_softmodem: Don't emit warning if server closes connection.
Currently, a warning is emitted when the TCP server closes
the connection, even though this is a legitimate and common
thing to happen.

This is changed to an appropriate verbose message, and we
also take care not to emit this log message multiple times,
since this could also be logged 4 or 5 times before hangup.