install: Fix linking regression in 21.1.0-rc1.
An upstream Asterisk regression caused linking
failure in builds using the latest RC. A hotfix
is incorporated to address that until a new
release candidate is released.

The ability to specify a specific Git commit hash
to compile a specific snapshot in time has also
been added, since this was used in debugging this issue.