install: Readd ALSA detection if needed.
Asterisk 21 removed chan_alsa, and with it
detection of the ALSA library. Other out-of-tree
modules (not in this repo) may still depend on
ALSA, so this allows the build system changes
to be reversed if needed so these modules can
still be built without modification.

This option does NOT patch chan_alsa back in.