restart: Improve safety of DAHDI restarts.
Stopping DAHDI while wanpipe is still running will cause the system
to lock up and crash. Add logic to the restart command to prevent
this, and add a "start" command with similar safety checks.
Also improve the restart command to be more robust and reliable
in case the DAHDI drivers need to be loaded in a particular order.