app_ccsa: Avoid queuing in absurd scenarios.
The CCSA() application previously happily queued
in many circumstances that were rather inappropriate
for either off-hook and/or callback queuing, in particular
when a destination couldn't be reached using any routes
and CCSA() would immediately offhook queue on a route.
Furthermore, since CCSA only pops items out of the queue
for call completion when a current call finishes, if there
are no calls using a route, queuing is pointless.

Add additional checks to avoid queuing altogether in these
scenarios and others like them.