res_alarmsystem: Add self-contained alarm system module.
res_alarmsystem is a module that can be used to implement
an alarm system within Asterisk. The basic functionality
is split into "client" and "server" roles, although
the "server" role is technically optional. This allows
the average home user with an Asterisk server at home
and off-site to have a professional-grade alarm system
with custom monitoring and reporting, all for the cost
of any sensors required.

WARNING: This module comes with ABSOLUTELY NO WARRANTY.
Use of it is solely at your own risk.

PHREAKSCRIPT-32 #close