DAHDI: Re-enable XPP drivers in 32-bit builds.
The XPP drivers were disabled from buildings on ~all 32-bit
architectures in 91b66b36d1c7a9d38512dd797b91a5b5208468c7,
as a temporary workaround to build failures with that architecture.

Apply an unmerged patch for 32-bit to fix the compilation issue so
that these drivers can be built in most (and possibly all?) cases
on 32-bit systems. The driver is patched and available again
on i386/i686 architectures (so 32-bit x86 systems).

The XPP driver was initially disabled for 32-bit ARM in
commit 6cedcea94a8290faf232778188db826ff53d86f8,
and for now I am leaving it disabled in that case
since I don't have a way to test that at the moment.
If this can be tested, this exception can be removed.