DAHDI: Successfully build against the latest Linux kernel versions. (#43)
To better test DAHDI against more recent kernels, compile the kernel
ourselves in the CI and build DAHDI against it, including any patches
necessary to successfully do so. This includes various kernel releases
in the 4, 5, and 6 major versions, all the way up to the current
mainline and next kernels. In the future, this should catch most
breakage as soon as it occurs.