install: Fix compilation issues with FreeBSD.
Make Asterisk compile successfully with FreeBSD. Since there is still
an issue with the GitHub CI not detecting the uuid package, the FreeBSD
build remains disabled in the CI, but should work if installing on a
regular FreeBSD system. Issues remain with actually running it; the
enclosed changes fix compilation issues only.