res_smdr_whozz: Add function to provide line state.
Add WHOZZ_LINE_STATE function, which returns whether a line is
currently off-hook, on-hook, or ringing. This is useful since
the DAHDI drivers do not provide a mechanism to distinguish
between off-hook and on-hook without disturbing the line (none
of the analog card registers provides this information). Thus,
the function can return information from the WHOZZ Calling?
hardware to provide this information within Asterisk.