install: Ensure that dahdi_tool is always built.
dahdi_tool requires the newt development packages, which
were installed when Asterisk was installed, but not when
DAHDI was installed. This meant dahdi_tool wouldn't be
built the first time DAHDI was compiled on a new system,
but it would be on future compilations. To ensure it's
always built, explicitly install the newt development
dependencies before compiling DAHDI Tools. Also, verify
that dahdi_tool has been built and fail the build if not.

PHREAKSCRIPT-55 #close