CI: Build DAHDI on Fedora-based distros.
Fedora-based distros were not previously built in the CI,
due to the provided kernel-devel packages not providing
headers that matched the running kernel.

Work around this by manually settings KSRC, in order to
use whatever kernel is available, allowing the build to
proceed.

Also add build for Alma Linux 9.5 (equivalent to RHEL 9.5),
so we can test the 9.5 minor version.