install: Add flag to auto-set KVERS. (#56)
For a little while, we've been manually setting KVERS in the
CI definitions (main.yml) in order to successfully build
against the provided kernel in the CI. However, the version
updates over time, so this breaks frequently. Since we just
want to match the version actually available, add a flag to
automatically set KVERS to the available kernel header version,
and just use that instead.