res_pbx_validate: Add module to find various dialplan bugs.
This adds a module that can automatically find many common dialplan bugs,
such as branches to nonexistent locations, unused dialplan contexts,
application argument misusage, etc.

Most of this code was written back in January 2022, but at the time the
patches were written to make changes to the PBX core. To avoid that,
all the logic has been moved to this single, out-of-tree module.

ASTERISK-29828