res_smdr_whozz: Add more detailed error logging.
Since initialization of the WHOZZ Calling? units can be finicky,
add more detailed error messages so the user has better visibility
into the actual issue, rather than just seeing a message that
the serial thread terminated.

Interestingly, as of now, it seems to be the case that the device
can be in a state where we cannot synchronize with it, but
establishing a serial connection using screen to the device works,
and then future attempts within res_smdr_whozz succeed normally.

While this hasn't been reconciled yet, the additional logging
should at least steer users in the right direction.