patches: Clone repo once instead of downloading patches individually.
Improve the performance of the patch phase of the install process by
doing a single shallow clone of the repo, rather than downloading
all needed files individually.

A side benefit of this is that we can now also force the CI to build
the version of a module in the checked out version of the repo, which
is useful for anything besides a commit to master (e.g. PR to master,
or a commit/PR to a different branch). Previously, they would all
download the master version, and to test a different build, the PR
would have to be temporarily modified to include the proper version.