install: Add option to install radio modules.
Add an option (--rpt) to install the radio modules (e.g. app_rpt, etc.).
These modules used to be in the Asterisk tree, but were removed in
commit https://github.com/asterisk/asterisk/commit/4585000039a1d94af44c1fa2ccb3b4490cdb869b
and a few other commits. The modules are currently being maintained in
a third-party repo, and this option pulls the files into the source tree.

If the test suite is being installed in conjunction with this option,
the radio tests are also added to the test suite.