res_smdr_whozz: Fix toggling settings not actually being set.
The SET_SETTING macro printed out that it was setting a setting,
but wasn't actually writing anything to the serial port. This
could result in the device not having the right settings and
not working right. In particular, after being unplugged for a
while and most likely losing its programming, the 'o' option
would not be set and we would get no CDR for outgoing calls.

PHREAKSCRIPT-64 #close