CI: Add builds that treat objtool failures as nonfatal.
Due to https://github.com/asterisk/dahdi-linux/issues/80, builds since
that commit have been failing and we are missing all other build issues
since this one. Add additional builds that revert this kernel commit for
now, so we can test for breakage since then. When this issue is fixed,
this can be undone.