app_callerid: Fix behavior for off-nominal cases.
Properly handle if Caller ID Name and/or Number are empty.
In particular, we need to check if the Caller ID fields are
valid before using them, to avoid copying uninitialized memory.
(Not that it would really make sense to send an empty Caller ID,
hence why this is an off-nominal case.)