app_softmodem: Add AMI command to retrieve list of active sessions.
Add an AMI command that will retrieve a list of active Softmodem sessions
(currently, all sessions are automatically tracked, though this could
be put behind an optional option if needed). This allows an external
program that is the target of a TCP connection initiated by the softmodem
to determine the identity of the caller, by using the TCP port to
correlate the session. One use case of this is to retrieve the Caller ID
of the modem caller from the TCP server, which could be used for
identification, authentication, or various other purposes.