res_alarmsystem: Add ALARMSYSTEM_STATE function.
Add a function that can be used to retrieve the current state
of the alarm system as a whole, rather than an individual
sensor. This allows us to determine from the dialplan if the
system is actively "in alarm" - just because a sensor is triggered
doesn't mean the system is, as the system may be in the temporarily
disarmed mode (e.g. to allow egress). This allows dialplan logic
to do certain things when an alarm is actively going off.