res_telos_1a2: Add workaround optimized for local FXS stations.
For lines served by analog Digium cards, regardless of the technology
behind the FXO port used, the fxodevice option (formerly just 'device')
does not work. This seems to be due to the high resistance of the line,
which prevents audio from passing through from an FXO port to the line.
In reality, the whole premise of using an FXO port the way we are is
completely non-standard and outside the scope of what should work, so
it's not surprising this can be undefined or varying behavior.

However, if the line itself is served by the system, as is likely in
the case of a Digium analog card, we don't need to use an FXO port in
order to bridge in music on hold; we can simply have Asterisk generate
it towards the caller. That is what the new 'fxsdevice' option does;
it queues hold on the channel along with the configured music on hold
class. This also results in better quality, since it eliminates an
additional analog connection and the volume is normal loudness.

UpgradeNote: The 'device' option has been renamed to 'fxodevice',
and this change is not backwards-compatible. The change was made
because, with the addition of the 'fxsdevice' option, having an
option simply named 'device' is somewhat ambiguous.

Additionally, the serial device no longer has a default, which ensures
the module does not load if it is not needed by the user.