res_smdr_whozz: Decline to load if 'device' setting not set.
This is a rather niche module that almost nobody will use;
therefore, comment this setting in the sample config and
decline to load if it isn't specified. This way, even with
the default sample configs installed, the module won't load
unless the config has been modified by the user. This
avoids a bunch of endless warnings in the console about
failing to write to the serial port when it likely hasn't
been configured for this use case.