Debian: Abort if apt-get update fails.
On systems installed from CD, if still present in sources.list
for whatever reason, and a disc isn't present, this can cause
both aptitude update and apt-get update to fail. This causes
Asterisk's install_prereq script to abort early (since it uses
set -e), as a result of which prereqs may not get installed,
eventually causing configure to fail.

Detect failure to update and abort immediately so we can
fail fast and alert the user to a problem with the package
manager configuration.

PHREAKSCRIPT-77 #close